Basic working dependency tree creator + visualizer, but major prob.
React UI to create nodes and edges + vis.js to render on update.

Already hit a major problem though. To add the interactivity I want
(such as clicking a node to edit its label, or clicking an edge to
delete it), I'll have to use the vis.js callbacks which totally go
around the react state.