2 tests: Click interaction + force layout by node value.
1. Click interaction: clicking a d3 node updates the react state
and re-renders smoothly. Here the text is simply changed to 'clicked',
but this means I know how to update the react state in a d3
interaction callback without mixing up whether react or d3 is responsible.

2. Use custom force functions to pull nodes to different heights based
on a value property. If I set that value to the # of nodes contained in
each nodes subtree I can layout my dependency DAG from top to bottom
in a way that makes sense. Had to increase the repellent force between
nodes to prevent them from forming a line down the center of the screen.