Maintain a 'selected' node in state.
Currently just coloring it to test how to do it, but the goal will
be to have a node editing UI form section that stays in sync with the
animation - so you can select a node in the graph, then edit that node's
text in the form.

One issue with managing the react state completely separately from the
d3 state (I'm not tracking the position of the nodes in the react
state) Is that when I click a node and update the selected state, the
whole graph re-positions. Wasn't a bother when adding nodes or links,
the disruption feels natural since the graph is changing, but when
simply selecting a node it feels weird for the whole layout to change.

Possible solution: Maintain a start position in the react state node
list, then use that in d3 as a starting point before relinquishing
all position control to d3.