Rewrite topological layout force.
Use the weights to target a height for each node. Logic like so:
1. Calculate node weights (roughly equal to number of dependencies)
2. Normalize weights to range [0,1]
3. Calculate target height by mapping from [0,1] -> [h * p, h * (1-p)]
   where h is height, and p is percentage padding.
4. Calculate node's speed towards target height proportional to distance
   from that target.

This works much better than the previous method because there is a
target. Before there were competing up/down forces that would take a
while to come to equilibrium. Also on rerendering there would always
be an initial jump because even if nodes were at a good position, there
was no consideration of being close to their equilibrium target - forces
were applied regardless.